moved reparsing of DB into seperate file so it won't conflict with the integration tests.
