+++ counterpartylib/test/unit_test.py
@@ -2,10 +2,10 @@
 import sys, os, time, tempfile
 import pytest
 import logging
-import util_test
-from util_test import CURR_DIR
-from fixtures.vectors import UNITTEST_VECTOR
-from fixtures.params import DEFAULT_PARAMS as DP
+from counterpartylib.test import util_test
+from counterpartylib.test.util_test import CURR_DIR
+from counterpartylib.test.fixtures.vectors import UNITTEST_VECTOR
+from counterpartylib.test.fixtures.params import DEFAULT_PARAMS as DP
 
 from counterpartylib.lib import (config, util, api, database, log)
 import server
